# Conflicts:
#	.gitignore
#	.travis.yml
#	build/CMakeLists.txt
#	files/samples/musicxml/Makefile
#	files/tests/Makefile
#	samples/Makefile
#	src/elements/versions.cpp
#	src/elements/versions.h
#	src/elements/xml.cpp
#	src/guido/xml2guidovisitor.cpp
#	src/guido/xml2guidovisitor.h
#	src/guido/xmlpart2guido.cpp
#	src/interface/libmusicxml.cpp
#	src/interface/libmusicxml.h
#	src/lib/ctree.h
#	src/lib/rational.cpp
#	src/operations/transposition.cpp
#	src/parser/xml.l
#	src/parser/xml.y
#	src/parser/xmllex.c++
#	src/parser/xmlparse.cpp
#	src/visitors/notevisitor.cpp
#	src/visitors/notevisitor.h
#	src/visitors/partsummary.h
#	src/visitors/xmlvisitor.cpp
#	validation/Makefile
#	validation/libmusicxmlversion.txt

+++ src/elements/xml.cpp
@@ -106,8 +106,8 @@ void xmlelement::setValue (float value)
 //______________________________________________________________________________
 long xmlelement::add (const Sxmlattribute& attr)
 { 
-  fAttributes.push_back(attr);
-  return fAttributes.size()-1;
+	fAttributes.push_back(attr);
+	return long(fAttributes.size()-1);
 }
 
 //______________________________________________________________________________
